requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
coverage==6.4.4
idna==3.3
requests==2.28.1
urllib3==1.26.12
python-dotenv~=0.21.0